+++ src/test/java/seedu/excolink/logic/commands/AddSubcomCommandTest.java
@@ -18,6 +18,7 @@ import javafx.collections.ObservableList;
 import seedu.excolink.commons.core.GuiSettings;
 import seedu.excolink.logic.commands.exceptions.CommandException;
 import seedu.excolink.model.Model;
+import seedu.excolink.model.ModelStub;
 import seedu.excolink.model.ReadOnlyExcoLink;
 import seedu.excolink.model.ReadOnlyUserPrefs;
 import seedu.excolink.model.person.Person;
@@ -78,91 +79,6 @@ class AddSubcomCommandTest {
         AddSubcomCommand command = new AddSubcomCommand(publicity);
         String expected = AddSubcomCommand.class.getCanonicalName() + "{toAdd=" + publicity + "}";
         assertEquals(expected, command.toString());
-    }
-
-    /**
-     * A default model stub that has all the methods failing.
-     */
-    private static class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getExcoLinkFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setExcoLinkFilePath(Path excoLinkFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setExcoLink(ReadOnlyExcoLink newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyExcoLink getExcoLink() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasSubcom(Subcom subcom) {
-            throw new AssertionError("This method should not be called");
-        }
-
-        @Override
-        public void addSubcom(Subcom subcom) {
-            throw new AssertionError("This method should not be called");
-        }
     }
 
     private static class ModelStubWithSubcom extends ModelStub {
